requirements.txt
numba==0.62.1
opencv-python==4.12.0.88
torch==2.9.0
pyyaml>=6.0